requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
pytest>=7.0.0
pytest-cov>=3.0.0
joblib>=1.1.0
